+++ js/settings-admin.js
@@ -23,9 +23,19 @@
  */
 
 window.addEventListener("DOMContentLoaded", function () {
-    $("#breezedark-enabled").change(function () {
+    $("#breezedark-theme-enabled").change(function () {
         $.post(OC.generateUrl("apps/breezedark/settings/admin"), {
             theme_enabled: this.checked ? 1 : 0,
+            theme_login_page: $("#breezedark-theme-login-page").prop("checked") ? 1 : 0,
+        });
+
+        $("#breezedark-theme-login-page").prop("disabled", !$("#breezedark-theme-enabled").prop("checked"))
+    });
+
+    $("#breezedark-theme-login-page").change(function () {
+        $.post(OC.generateUrl("apps/breezedark/settings/admin"), {
+            theme_login_page: this.checked ? 1 : 0,
+            theme_enabled: $("#breezedark-theme-enabled").prop("checked") ? 1 : 0,
         });
     });
 });